Integrated PDF scale controls into menu bar
