+++ src/apps/annotation-text/Menubar/PDFControls.css
@@ -1,0 +1,26 @@
+.anno-menubar .pdf-scale-controls {
+  align-items: center;
+  display: flex;
+}
+
+.anno-menubar .pdf-scale-controls button.select-trigger {
+  border-radius: var(--border-radius);
+  width: auto;
+  min-width: 90px;
+  padding: 0 0.8em;
+}
+
+.anno-menubar .pdf-scale-controls .custom-scale > span:first-child {
+  color: transparent;
+}
+
+.anno-menubar  .pdf-scale-controls .custom-scale > span:last-child {
+  align-items: center;
+  display: flex;
+  height: 100%;
+  justify-content: center;
+  left: 0;
+  position: absolute;
+  top: 0;
+  width: 100%;
+}
