--- src/components/AnnotationDesktop/DeprecatedDocumentMenu/DocumentMenu.css
@@ -1,83 +1,0 @@
-.document-menu {
-  align-items: center;
-  border-radius: 0 0 12px 0;
-  border-width: 0 0 1px 1px;
-  box-sizing: border-box;
-  display: flex;
-  gap: 0.5em;
-  height: 48px;
-  padding: 0 20px 0 10px;
-  pointer-events: all;
-}
-
-.document-menu h1 {
-  font-size: var(--font-tiny);
-  font-family: Inter;
-  font-weight: 500;
-  margin: 0;
-}
-
-.document-menu .assignment-icon {
-  align-items: center;
-  background-color: var(--font-dark);
-  border-radius: 50%;
-  color: #fff;
-  display: inline-flex;
-  height: 28px;
-  justify-content: center;
-  width: 28px;
-}
-
-.document-menu a {
-  color: var(--font-dark);
-}
-
-.document-menu a svg {
-  display: block;
-}
-
-.document-menu span {
-  color: var(--font-light);
-}
-
-.document-menu button {
-  border-radius: 50%;
-  height: 36px;
-  vertical-align: bottom;
-  width: 36px;
-}
-
-.document-menu .pdf-scale-controls {
-  align-items: center;
-  display: flex;
-}
-
-.document-menu .pdf-scale-controls button.select-trigger {
-  border-radius: var(--border-radius);
-  width: auto;
-  min-width: 90px;
-  padding: 0 0.8em;
-}
-
-.document-menu .pdf-scale-controls .custom-scale > span:first-child {
-  color: transparent;
-}
-
-.document-menu .pdf-scale-controls .custom-scale > span:last-child {
-  align-items: center;
-  display: flex;
-  height: 100%;
-  justify-content: center;
-  left: 0;
-  position: absolute;
-  top: 0;
-  width: 100%;
-}
-
-.document-menu button svg {
-  color: var(--font-dark);
-}
-
-.document-menu :is(button, a.button) + :is(button, a.button) {
-  margin-left: 0;
-}
