--- src/components/AnnotationDesktop/DeprecatedDocumentMenu/index.ts
@@ -1,1 +1,0 @@
-export * from './DocumentMenu';
