+++ src/themes/default/annotation/index.css
@@ -62,6 +62,10 @@
   z-index: 2;
 }
 
+.anno-menubar h1 {
+  margin-right: 8px;
+}
+
 .anno-menubar .anno-menubar-left {
   align-items: center;
   border-radius: 0 0 12px 0;
@@ -76,7 +80,6 @@
   font-size: var(--font-tiny);
   font-family: Inter;
   font-weight: 500;
-  margin: 0;
 }
 
 .anno-menubar .anno-menubar-left .assignment-icon {
